pyproject.toml
[project]
name = "diet-coach"
version = "0.1.0"
description = "Diet Coach - AI agent tryout playground."
readme = "README.md"
authors = [
    { name = "Matze", email = "matthias.teloeken@posteo.de" }
]
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = [
    "mistralai>=1.12.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

#[project.scripts]
#dwd-mcp-server = "dwd_mcp_server.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--import-mode=importlib"]

[tool.ruff]
line-length = 120
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
select = [
    "A",  # flake8-builtins
    "ANN",  # mypy
    "ARG",  # flake8-unused-arguments
    "B",  # Bugbear
    "C4",  # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "E",  # pycodestyle
    "F",  # Pyflakes
    "FLY",  # flynt
    "FA",  # flake8-fututre-annotations
    "I",  # isort
    "PTH",  # flake8-use-pathlib
    "SIM",  # flake8-simplify
    # "T20",  # flake8-print
    "TCH",  # flake8-type-checking
    "UP",  # pyupgrade
    "W",  # pycodestyle
    "C90",  # mccabe
]

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
typeCheckingMode = "basic"
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"
